persistencia en settings funcionando
