+++ .idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/el_ma/Desktop/ORT/Cuarto-cuatrimestre/TP3/proyectos/tp-final/rick-and-morty-tp3/app/src/main/res/menu/drawer_menu.xml" value="0.36770833333333336" />
+        <entry key="..\:/Users/el_ma/Desktop/ORT/Cuarto-cuatrimestre/TP3/proyectos/tp-final/rick-and-morty-tp3/app/src/main/res/xml/user_setting.xml" value="0.3596014492753623" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.2155" />
         <entry key="app/src/main/res/drawable/hamburguer.xml" value="0.2435" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.2155" />
