+++ app/src/main/res/navigation/nav_graph.xml
@@ -34,11 +34,19 @@
         <argument
             android:name="username"
             app:argType="string" />
+        <action
+            android:id="@+id/action_homeFragment_to_settingsActivity"
+            app:destination="@id/settingsFragment" />
     </fragment>
+
+    <activity
+        android:id="@+id/settingsActivity"
+        android:name="com.example.rick_and_morty_tp3.MySettingsActivity"
+        android:label="SettingsActivity" />
 
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="com.example.rick_and_morty_tp3.fragments.SettingsFragment"
+        android:name="com.example.rick_and_morty_tp3.MySettingsActivity$SettingsFragment"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
